requirements.txt
Flask==2.3.2
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.1
seaborn==0.12.2
requests==2.31.0
python-dotenv==1.0.0
mysql-connector-python==8.0.33
datetime==4.7.0
wordcloud===1.8.2

